% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_avemax.R
\name{check_avemax}
\alias{check_avemax}
\title{Check averages and maxima}
\usage{
check_avemax(dat, flag)
}
\arguments{
\item{dat}{Data frame to be tested.}

\item{flag}{Do you want to check if ave is greater than max ("flag") or if
there are missing values when either ave or max is present ("missing")}
}
\description{
Checks that averages are not larger than maxima.
}
\author{
Kara Woo
}

