% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_dates.R
\name{check_dates}
\alias{check_dates}
\title{Check dates}
\usage{
check_dates(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Make sure that start date is before end date.
}
\author{
Kara Woo
}

