% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_depths.R
\name{check_depths}
\alias{check_depths}
\title{Check sample depth and photic depth}
\usage{
check_depths(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
The sampledepth field should be the depth of the samples used in
aggregation, not the maximum depth sampled (i.e. if sampling was done
below the photic zone). Therefore, sampledepth should not be greater than
photicdepth, or not by much (a few meters is probably ok). This function
returns a data frame of observations where the sample depth provided is
greater than the photic depth.
}
\author{
Kara Woo
}

