% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_doc.R
\name{check_doc}
\alias{check_doc}
\title{Check DOC values}
\usage{
check_doc(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
We originally asked for DOC data in units of ug/l, however mg/l was provided
by most researchers and is a more standard unit. If DOC is in the hundreds
or thousands, it is likely in ug/l and we should follow up with researchers
and likely convert their data.
}
\author{
Kara Wo
}

