% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_fields.R
\name{check_fields}
\alias{check_fields}
\title{Check field names}
\usage{
check_fields(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Match column names to data template fields. Returns a list with two elements:
column names in the data frame that are not present in the template, and
fields in the template that are missing from the data frame. Result is NULL
if all fields match.
}
\author{
Kara Woo
}

