% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/check_dates.R
\name{check_iceduration_iceoff}
\alias{check_iceduration_iceoff}
\title{Check ice duration column during iceoff period}
\usage{
check_iceduration_iceoff(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Ice duration should be zero during iceoff.
}
\author{
Kara Woo
}

