% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_dates.R
\name{check_iceduration_length}
\alias{check_iceduration_length}
\title{Check ice duration}
\usage{
check_iceduration_length(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Make sure ice duration is not shorter than aggregation period in winter.
}
\author{
Kara Woo
}

