% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_neg.R
\name{check_neg}
\alias{check_neg}
\title{Check for negative values}
\usage{
check_neg(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Checks that numeric columns are never negative. These columns (with the
exception of stationlat, stationlong, and airtemp which can be negative)
should not contain values less than zero.
}
\author{
Kara Woo
}

