% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_nitro.R
\name{check_nitro}
\alias{check_nitro}
\title{Check nitrogen data}
\usage{
check_nitro(dat)
}
\arguments{
\item{dat}{Data frame to be tested}
}
\description{
Check order of magnitude of total nitrogen and total dissolved
nitrogen. This data should be in units of μg/l; if values are <=15 check
that they aren't in mg/l, and if so convert. Also checks if dissolved
nitrogen values are larger than total nitrogen and returns a data frame of
any such values.
}
\author{
Kara Woo
}

