% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/check_num.R
\name{check_num}
\alias{check_num}
\title{Check numeric fields}
\usage{
check_num(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Certain fields should be numeric; check that they are. Returns a vector of
column names of fields that should be numeric but aren't.
}
\author{
Kara Woo
}

