% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/check_dates.R
\name{check_periodn}
\alias{check_periodn}
\title{Check periodn}
\usage{
check_periodn(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Should not be smaller than the number of days in the aggregation period
(this is confusing in the instructions).
}
\author{
Kara Woo
}

