% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_phytomass.R
\name{check_phytomass}
\alias{check_phytomass}
\title{Check phytoplankton biomass}
\usage{
check_phytomass(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
We originally asked for phytoplankton biomass in units of ug dry weight /
liter, however after talking with researchers we determined that biovolume
is a better unit. This function throws a warning if biomass data is present
for phytoplankton, ciliates, or heterotrophic nanoflagellates so that we can
follow up with the researcher about converting the units.
}
\author{
Kara Woo
}

