% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_props.R
\name{check_props}
\alias{check_props}
\title{Check proportions}
\usage{
check_props(dat)
}
\arguments{
\item{dat}{Data frame to be tested}
}
\description{
Proportions of different zooplankton and phytoplankton taxa should sum to 1.
Returns a list of data where proportions do not sum to 1 for phytoplankton
and zooplankton taxa. If all proportions sum to 1, no list is returned.
}
\author{
Kara Woo
}

