% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_repeats.R
\name{check_repeats}
\alias{check_repeats}
\title{Check repeats}
\usage{
check_repeats(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
There shouldn't be more than one observation from a single site in a single
year and season. This function checks for repeats in both stationname and
latitude/longitude. Note that if there are multiple stations with different
coordinates and no names (i.e. NA in the stationname field), these will be
shown as repeats.
}
\author{
Kara Woo
}

