% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/check_values.R
\name{check_values}
\alias{check_values}
\title{Check defined values}
\usage{
check_values(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Certain fields can only have certain values. This function checks that
the following fields contain only allowed values:
- season ("iceon" or "iceoff")
- multiplestations ("yes" or "no")
- startmonth and endmonth (three letter abbreviation)
- sampletype ("in situ" or "remote sensed")
- fadata ("no", "proportional", or "concentrations")
- gutdata ("yes" or "no")
- bensubstrate (NA or "organic", "silt", "sand", "rock", "mixed")
}
\author{
Kara Woo
}

