% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/check_avemax.R
\name{compNA}
\alias{compNA}
\title{Compare ave and max for NAs}
\usage{
compNA(a, b)
}
\arguments{
\item{a}{first argument}

\item{b}{second argument}
}
\description{
Returns TRUE if both or neither argument is NA; returns FALSE if one is NA
but the other is not.
}
\author{
Kara Woo
}

