% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_num.R
\name{convert_num}
\alias{convert_num}
\title{Convert fields to numeric with meaningful warnings}
\usage{
convert_num(x, y, coerce = FALSE)
}
\arguments{
\item{x}{Column to be converted}

\item{y}{Data frame}

\item{coerce}{Should NAs be introduced if conversion to numeric fails?}
}
\description{
Convert character fields to numeric, and if conversion fails give an
informative warning.
}
\author{
Kara Woo
}

