% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_fields.R
\name{field_match}
\alias{field_match}
\title{Field match}
\usage{
field_match(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Checks field names and returns TRUE if all column names match template
names, FALSE if names do not match.
}
\author{
Kara Woo
}

