% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/check_num.R
\name{fix_num}
\alias{fix_num}
\title{Fix non-numeric fields}
\usage{
fix_num(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Convert fields from character to numeric.
}
\author{
Kara Woo
}

