% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_props.R
\name{prop_sums}
\alias{prop_sums}
\title{Calculate proportion sums}
\usage{
prop_sums(x, cols)
}
\arguments{
\item{x}{Data frame to be checked}

\item{cols}{Columns which should sum to one in each row.}
}
\description{
Sum proportions and, if any do not sum to 1, return a data frame of some
identifiable columns of the data (year, season, lakename, station name,
lat/long) and the proportion sum.
}

