% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/t_icedata.R
\name{t_icedata}
\alias{t_icedata}
\title{Transform data from horizontal to vertical.}
\usage{
t_icedata(dat)
}
\arguments{
\item{dat}{Data frame to be tested. Should be in the format of the data
template (including the colums 'datacategory', 'dataclass', 'fieldname',
and 'fielddescription..follow.format.of.quoted.texts.'). Import with
read.csv("filename.csv", stringsAsFactors = FALSE).}
}
\description{
Converts data from horizontal (matching the data template) to vertical (with
field names as column headers) for use with the rest of the functions in
this package.
}
\author{
Kara Woo
}

