% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/check_avemax.R
\name{check_avemax}
\alias{check_avemax}
\title{Check averages and maxima}
\usage{
check_avemax(dat, flag)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
Checks that averages are not larger than maxima.
}
\author{
Kara Woo
}

