% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/check_repeats.R
\name{check_repeats}
\alias{check_repeats}
\title{Check repeats}
\usage{
check_repeats(dat)
}
\arguments{
\item{dat}{Data frame to be tested.}
}
\description{
There shouldn't be more than one observation from a single site in a single
year and season.
}
\author{
Kara Woo
}

